% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo2batchimport.R
\name{photo2batchimport}
\alias{photo2batchimport}
\title{Create the "batch import" table for Amphibian and Reptile Wildbook}
\usage{
photo2batchimport(
  BatchImportdf = BatchImportdf,
  submitterID = "Votre Nom",
  MarkedID = F
)
}
\arguments{
\item{BatchImportdf}{A dataset to be processed}

\item{submitterID}{Name of the person submitting the data}
}
\value{
The table to upload on ARW Batch Import
}
\description{
Create the "batch import" table for Amphibian and Reptile Wildbook
}
\examples{
\dontrun{
new_dataset <- photo2batchimport(dataset)
}
}
